+++ pkgs/intl4x/lib/src/list_format/list_format_impl.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:meta/meta.dart' show ResourceIdentifier;
+
 import '../../ecma_policy.dart';
 import '../data.dart';
 import '../ecma/ecma_policy.dart';
@@ -20,7 +22,8 @@ abstract class ListFormatImpl {
 
   String formatImpl(List<String> list);
 
-  factory ListFormatImpl.build(
+  @ResourceIdentifier('ListFormat')
+  static ListFormatImpl build(
     Locale locales,
     Data data,
     ListFormatOptions options,